chore: move tests back to jest
