+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["./flipt-engine-ffi", "./flipt-client-browser", "./flipt-evaluation"]
+members = ["./flipt-engine-ffi", "./flipt-engine-wasm", "./flipt-evaluation"]
 
 [profile.release]
 strip = true
