+++ flipt-client-browser/package.json
@@ -17,14 +17,16 @@
   "scripts": {
     "fmt": "prettier --config .prettierrc 'lib/**/*.ts' --write",
     "lint": "prettier --check lib/**/*.ts",
-    "test": "jasmine",
-    "build:typescript": "tsc",
-    "build:wasm": "wasm-pack build --target web --out-dir ./pkg",
-    "build": "npm run build:wasm && npm run build:typescript"
+    "test": "jest",
+    "build": "tsc"
   },
   "devDependencies": {
+    "@babel/core": "^7.24.4",
+    "@babel/preset-env": "^7.24.4",
+    "@types/jest": "^29.5.12",
     "@wasm-tool/wasm-pack-plugin": "^1.7.0",
-    "jasmine": "^5.1.0",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.2",
     "typescript": "^5.4.3"
   }
 }
