--- flipt-client-browser/spec/support/jasmine.json
@@ -1,9 +1,0 @@
-{
-  "spec_dir": "spec",
-  "spec_files": ["**/*[sS]pec.?(m)js"],
-  "helpers": ["helpers/**/*.?(m)js"],
-  "env": {
-    "stopSpecOnExpectationFailure": false,
-    "random": true
-  }
-}
