+++ flipt-client-node/jest.config.js
@@ -1,6 +1,4 @@
 module.exports = {
-  roots: ['<rootDir>/src'],
-  testMatch: ['**/?(*.)+(spec|test).+(ts)'],
   transform: {
     '^.+\\.(ts)$': 'ts-jest'
   }
