+++ flipt-client-node/tsconfig.json
@@ -7,6 +7,12 @@
     "strict": true,
     "esModuleInterop": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist/**/*"]
+  "include": [
+    "src/**/*",
+    "__tests__"
+  ],
+  "exclude": [
+    "node_modules",
+    "dist/**/*"
+  ]
 }
